split out the UserMixin to make picking & choosing easier
I wanted to be able to have a user object for PAI that was very
different attribute-wise from the auth user.  But, the user
would be put in groups and have permissions etc.  So I am
trying to break out that logic so the relationship functionality
can be easily mixed in without relying on other attributes of the
user.